Isolate actionable types
