--- src/models/_actionable/index.ts
@@ -1,3 +1,0 @@
-export * from './creatable-actionable';
-export * from './raw-creatable-actionable';
-export * from './to-raw-creatable-actionable';
