+++ src/models/actionable/creatable-actionable.ts
@@ -1,6 +1,14 @@
+/*************************************************************************
+ * Copyright 2020 Gravwell, Inc. All rights reserved.
+ * Contact: <legal@gravwell.io>
+ *
+ * This software may be modified and distributed under the terms of the
+ * MIT license. See the LICENSE file for details.
+ **************************************************************************/
+
 import { ConditionalPartial, PartialProps } from '../../functions/utils';
 import { NumericID } from '../../value-objects';
-import { ActionableAction, ActionableTrigger } from '../actionable';
+import { ActionableAction, ActionableTrigger } from './actionable';
 
 export interface CreatableActionable {
 	userID?: NumericID;
