+++ src/models/actionable/raw-creatable-actionable.ts
@@ -1,5 +1,13 @@
+/*************************************************************************
+ * Copyright 2020 Gravwell, Inc. All rights reserved.
+ * Contact: <legal@gravwell.io>
+ *
+ * This software may be modified and distributed under the terms of the
+ * MIT license. See the LICENSE file for details.
+ **************************************************************************/
+
 import { RawNumericID, RawUUID } from '../../value-objects';
-import { RawActionableAction, RawActionableTrigger } from '../actionable';
+import { RawActionableAction, RawActionableTrigger } from './raw-actionable';
 
 export interface RawCreatableActionable {
 	// ?Question: Why do we allow that field to be set? gravwell/gravwell#2318 nº 3
