+++ src/models/actionable/to-raw-creatable-actionable.ts
@@ -1,6 +1,15 @@
+/*************************************************************************
+ * Copyright 2020 Gravwell, Inc. All rights reserved.
+ * Contact: <legal@gravwell.io>
+ *
+ * This software may be modified and distributed under the terms of the
+ * MIT license. See the LICENSE file for details.
+ **************************************************************************/
+
 import { omitUndefinedShallow } from '../../functions/utils';
 import { toRawNumericID } from '../../value-objects';
-import { ActionableAction, toRawActionableAction } from '../actionable';
+import { toRawActionableAction, toRawActionableTrigger } from '../actionable';
+import { ActionableAction } from './actionable';
 import { CreatableActionable } from './creatable-actionable';
 import { RawCreatableActionable } from './raw-creatable-actionable';
