+++ src/models/index.ts
@@ -35,4 +35,3 @@ export * from './timeframe';
 export * from './user';
 export * from './user-preferences';
 export * from './version';
-export * from './_actionable';
